# ns-topup/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NS Games Topup API",
        "description": "API untuk platform topup game, pulsa, dan e-wallet dengan integrasi Digiflazz dan Midtrans",
        "contact": {
            "name": "NS Games Support",
            "email": "admin@nsgames.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Laravel Development Server"
        },
        {
            "url": "http://localhost/NS-topupgames/ns-topup/public",
            "description": "Laragon/XAMPP Server"
        }
    ],
    "paths": {
...
        "/api/customer/profile": {
            "get": {
                "tags": [
                    "Customer - Profile"
                ],
                "summary": "Get customer profile",
                "operationId": "8b507affe0fcad8fb9934b84dcc8d113",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "081234567890"
                                        },
                                        "balance": {
                                            "type": "number",
                                            "example": 50000
                                        },
                                        "level": {
                                            "type": "string",
                                            "example": "Bronze"
                                        },
                                        "total_orders": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-01-15T08:30:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Customer - Profile"
                ],
                "summary": "Update customer profile",
                "operationId": "a1adb820728515a51c47bdfa63a3345a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "081234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile updated successfully"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/customer/topup-balance": {
            "post": {
                "tags": [
                    "Customer - Profile"
                ],
                "summary": "Top up customer balance",
                "operationId": "2161fc8f726338233c6acf7e7d79a63c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "payment_method"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "example": 100000
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "qris"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Top up request created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "topup_id": {
                                            "type": "string",
                                            "example": "TOP-20241201-001"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "example": 100000
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "waiting_payment"
                                        },
                                        "payment_url": {
                                            "type": "string",
                                            "example": "https://payment.example.com/topup/123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}